Add files via upload
Experts on the Wall Street might predict that the evaluation of a particular stock might balloon or plummet. But now it is a common belief that the financial experts rely greatly on analytical tools and algorithms to predict the future of a stock price by considering a myriad of variables and then making a prediction of the stock. In this highly digitized world where data is the new gold, prediction software’s and algorithms are getting more efficient and reliable. Investors and conglomerates that earn their millions by investing into stocks, can maximize their profit margins by investing in the right stock. As Norman Augustine said- ‘If stock market experts were so expert, they would be buying stocks and not giving people selling advice’, our algorithm will make statistical analysis of the selected stock, create visualizations of the trend and help select a stock for investment easier by making predictions by analyzing the real time data.